Move Sketch tests into the module as unit tests
